feat: 에러 메시지 출력  interceptor로 통합

+++ lib/common/dio/dio.dart
@@ -39,7 +39,9 @@ class CustomInterceptor extends Interceptor {
   @override
   void onError(DioError err, ErrorInterceptorHandler handler) async {
     print('[ERR] [${err.requestOptions.method}] ${err.requestOptions.uri}');
-
+    print(err.response!.data);
+    print(err.error);
+    print(err.message);
     final isStatus403 = err.response?.statusCode == 403;
     final isPathRefresh = err.requestOptions.path ==
         'https://mohim-api.com/api/v1/auth/refresh-token';
@@ -49,12 +51,17 @@ class CustomInterceptor extends Interceptor {
         err.type == DioErrorType.sendTimeout ||
         err.type == DioErrorType.other) {
       Get.snackbar('네트워크 연결오류', '네트워크를 연결해주세요');
-    }
-    // 나머지 오류에는 재로그인 요청
-    else {
+    } else if (err.response?.statusCode == 403) {
       await storage.deleteAll();
       Get.snackbar('재로그인 필요', '모힘을 찾아주신지 오래되셨네요! 다시 로그인 해주세요');
       Get.offAll(AuthBranchScreen());
+    }
+    // 나머지 오류에는 재로그인 요청
+    else {
+      Get.rawSnackbar(
+        message: err.response!.data['message'].toString(),
+        animationDuration: Duration(milliseconds: 400),
+      );
     }
 
     return handler.next(err);
